Related: tdf#113076 vcl win DirectWrite: handle stretched text
Commit b4f887aa3b838df2df2f4a28baf74a71f07ce3b3 (tdf#103831, tdf#100986:
Force using GDI when needed, 2017-03-03) noted that the DirectWrite text
renderer doesn't support stretched text, add support for this now by
setting a DirectWrite transform matrix that only does horizontal
scaling.

With this, tdf#113076 is kept fixed, but at the same time manually
stretched text keeps working. Previously the glyphs of the text had the
correct size and position, but the glyphs themselves where not streched,
but simply aligned to the left.

Change-Id: I8fe8e74d3edc0d71ed2f16fcce66c6f5009ed264
Reviewed-on: https://gerrit.libreoffice.org/71245
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins