tdf#124953: Use rangelist's combined range top-left address...
as origin address in the conditional format formula on xlsx export.
Excel seems to get confused if anything else is supplied as the
origin in the formula.

For example, before this patch, for a condfmt range over the
range-list [A3:C5 E1:F2], the origin address used in the formula
(for text search type entries) is A3.

<conditionalFormatting sqref="A3:C5 E1:F2">
   <cfRule type="containsText" dxfId="0" priority="1" operator="containsText" text="ABC">
       <formula>NOT(ISERROR(SEARCH("ABC",A3)))</formula>
   </cfRule>
</conditionalFormatting>

In this patch we use the top-left cell address(A1) of the combined range "A1:F5" as
the origin address.

       <formula>NOT(ISERROR(SEARCH("ABC",A1)))</formula>

Change-Id: If08a859bc361f925148ff463758d03ebbc41c0ac
Reviewed-on: https://gerrit.libreoffice.org/71312
Tested-by: Jenkins
Reviewed-by: Andras Timar <andras.timar@collabora.com>