Just apply Asian kerning to unmodified glyphs
IMHO the old code was doing the opposite of the comment. But the
comment makes more sense then the code, so adjust the code
accordingly.

OTOH just this code was added in commit f89d70de85a6 ("INTEGRATION:
CWS vcl27 (1.30.48); FILE MERGED 2004/09/09 12:33:53 hdu
1.30.48.1: #116489# only apply asian kerning on default-formatted
text"). Maybe still a bug? Can't check the bug tracker, as this is
pre-Apache, and so can't verify the old bug :-(

Change-Id: I550d1524d3e82f7e853820d3e5ddab8cc62228f9
Reviewed-on: https://gerrit.libreoffice.org/71441
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>