tdf#123285 sw_redlinehide: fix SwAutoFormat::DelMoreLinesBlanks()
The "Delete spaces and tabs at end and start of line" AutoFormat
option was deleting the wrong characters: on the 2nd line, it deletes
the spaces as it should, but on subsequent lines the selection is off by
the number of characters that have already been removed.

The range indexes are now integers, so not automatically corrected on
deletions like the SwIndex they were before; in this case, where only
plain characters are deleted, it should suffice to iterate the ranges in
reverse order to avoid the problem.

(regression from f6a5367ee5a85c75eb324ad9b9e3cafb86741753)

Change-Id: Ie8e08c00803d94e700cd5b64f29edb0b841eda2b
Reviewed-on: https://gerrit.libreoffice.org/71686
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>