Resolves: tdf#124605 ditch "if operand 1 is Empty, result is operand 2"
It has been like that since the initial import but is utter
nonsense. It might had (doubtfully) served some early StarBasic
compatibility quirk, but is wrong and was implementation defined
buggy behaviour. The Option VBASupport 1 even explicitly disabled
it.

In future we may want to implement the VBA Nothing value for an
SbxEMPTY at least for boolean operators, but this for Calc user
defined macro functions might even need a distinguished
SbxEMPTYCELL or such. Or an explicit SbxNOTHING.

Change-Id: I28919d982d0e60b9b840a12271dc717effa59662
Reviewed-on: https://gerrit.libreoffice.org/71701
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins