tdf#122892 tdf#124366 sw: fix loop in SwTextFrame::CalcFollow()
... differently; commit 0be58e5a38cf48c05518ec8c515da37b0749be27
was partially wrong because if SetPrepWidows() isn't called and the
master and follow have fewer lines than the sum of orphans and
widows, then the frames must be merged, i.e., master must move
forward to the next page and follow must be deleted, but that only
happens with SetPrepWidows().

However if the SwTextFrame::PrepWidows() determines that no lines can be
moved from the master, SwTextFrame::CalcPreps() will grow the master
to force it to move to the next page, and then it sets SetWidow() too,
so check this flag to break the loop in SwTextFrame::CalcFollow().

Another question is what is the point of the "very cheesy" follow
formatting fall-back, which might be removable.

Change-Id: I0769a5a4f13ca4c95c2229a496207809d05576f9
Reviewed-on: https://gerrit.libreoffice.org/71866
Tested-by: Jenkins
Reviewed-by: Michael Stahl <Michael.Stahl@cib.de>