tdf#124594 DOCX filter: don't extend margins from effects for rotated shapes
Regression from commit d71edc4ddae341f6cb468e7dcf59b7252857a433 (DOCX
filter: effect extent should be part of the margin, 2014-12-04), the
problem was that extending margins as-is based on the effect extent
values only work correctly in case of non-rotated shapes.

For example, with 90 degree clockwise rotation the top effect extent
should extend the right margin, etc. Fix the bug by limiting this
extension to the non-rotated scenario.

Test the behavior at a layout level, so in case later the effect extent
feature is implemented, it won't be necessary to adjust the test.

Change-Id: I97271bbb7c079951980b436cb8d8e5e54eeead55
Reviewed-on: https://gerrit.libreoffice.org/71878
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>