tdf#129071 Qt5 set file picker parent widget
If the XInitialization::initialize has a parent option, use that
system window ID to find the parent Qt5Frame and set this as the
parent of the file picker. This way the file picker doesn't show
up as a separate window in the KDE task bar and get the proper
icon. Just setting it transient to the parent is not enough.

This also includes the terminate listener handling, so an open
file picker won't prevent LO to shut down gracefully (which is
handled independent from the document modified state).

Change-Id: I62aef532e739e7fea9fe8a4b4e6c0e205faae1d9
Reviewed-on: https://gerrit.libreoffice.org/74544
Tested-by: Jenkins
Reviewed-by: Jan-Marek Glogowski <glogow@fbihome.de>