tdf#125591 DOC import: lazy-load metafiles with explicit size
Regression from commit 7e6acdea18b34d770b6ee01f7be655c0a170e3b8
(tdf#125281 DOC import: fix size of lazy-loaded metafiles, 2019-05-27),
the problem is that setting the preferred size of a Graphic swaps it in.

Avoid this by extending ImportUnloadedGraphic(): if a size hint is
provided, then that will be used instead of info from the graphic
descriptor (which is usually only meaningful for bitmaps).

This way we maintain the correct size and we're back to lazy-loading
metafiles from binary MSO files as well.

Change-Id: Ide12d12166110e98ea47b5347dd24fb203b22da3
Reviewed-on: https://gerrit.libreoffice.org/73798
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins