Related: tdf#113541 use reference syntax of the document being edited
When creating a formula referencing an external document by
travelling (clicking in) to that document, the reference was
created using the reference syntax of the referenced document,
which may be different if the syntax was changed for only one of
the documents after being opened. Use the reference syntax of the
document being edited instead so the created reference is correct
and can be compiled.

Change-Id: I07d0ce9cd911ac8e62576b4cdabdfd2569336fbf
Reviewed-on: https://gerrit.libreoffice.org/73983
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Jenkins